Fix Strage Nav Borders
Fixed the top navigation bar wrapping onto the secons line. We have
just fored it into the responsive mode (Menu button to reveal) on
screens smaller than 1063px. To achive this we needed to add an Admin
class to the body and then target what we need. The navigation border
styling was also applying to other elements because it used .nav used
a slight more speciofic class .main-nav to stop this from happening.