Fix IE Cross Browser Issues
Fixed a few bugs that appear in IE8. Header had an unexpected 20px
padding when > ~760px (like its in mobile sizes) and the focus style
was too dark and made the text unreadable.

The header had a 20px spacing to the side of it. This looked to be due
to the @media polyfil we use, as its a js polyfil and we call the
@media in dynamic base then override in the child theme, IE isn't
overriding it as it looks like the javsript is rendered/injected
last (as you would kinda expect).