Fix Validation errors
Fixed various validation errors ande removed unessarry langauge
attribute on script tag.

The only fix that may cause problems is removing cellpadding cellspacing
border and border attributes from the tables. It doesn't seemt to be a
probelm so far, but if the layout is changed on any reports what these
attrbutes achieve needs to be done through CSS.